backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
dill==0.4.0
fastapi==0.118.0
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
idna==3.10
joblib==1.5.2
keras==3.12.0
libclang==18.1.1
llvmlite==0.45.1
markdown==3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
ml-dtypes==0.5.3
namex==0.1.0
numba==0.62.1
numpy==2.3.3
opt-einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.0
psycopg2-binary==2.9.10
pydantic==2.12.0
pydantic-core==2.41.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
requests==2.32.5
rich==14.2.0
scikit-learn==1.6.1
scipy==1.16.3
setuptools==80.9.0
shap==0.49.1
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
sqlalchemy==2.0.43
starlette==0.48.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
